Remove duplicate tests
